Added the following description of problems with inflation to Pg 36: In this thesis inflation is taken to be the mechanism by which inhomogeneities in matter are seeded and the horizon and flatness problems of the Big Bang are solved. However, the inflationary paradigm is not without its own challenges.
Chief amongst these is the lack of a unique underlying theory. Many high energy theories have been
shown to produce an inflationary phase. Often, however, these require a great deal of fine-tuning
in order to produce a sufficient number of e-foldings of inflation. Lack of knowledge about the
governing physics at high energy scales hampers our understanding of the cause of inflation and
undermines any analysis of the generic nature of the initial conditions required.

The overall duration of inflation is also unknown. Observations only require that currently
observable scales were previously inside the horizon. Thus the onset of inflation is not
constrained and could occur far in the past. However, allowing such a long inflationary period
typically increases the fine-tuning necessary and can lead to other issues. 

There are further problems with the inflationary paradigm, including the lack of an explanation for
how energy in the inflaton field is transferred to the other constituent parts of the universe, and
indeed the fact that no scalar field has yet been directly observed.

We will continue to employ the inflationary paradigm in this thesis but it is important to
acknowledge that some challenges remain to be overcome.